# Improve performance of `Similarity::calculateVectors`